+++ eds_scikit/biology/__init__.py
@@ -1,6 +1,7 @@
-from .viz import plot_biology_summary
+from .viz import plot_biology_summary, measurement_values_summary
 from .cleaning import bioclean
 from .utils.process_concepts import ConceptsSet, fetch_all_concepts_set
 from .utils.config import register_configs
+from .utils.prepare_measurement import prepare_measurement_table
 
 register_configs()
