+++ eds_scikit/biology/utils/prepare_measurement.py
@@ -4,7 +4,7 @@ from eds_scikit.utils.framework import to
 
 from eds_scikit.biology.utils.check_data import check_data_and_select_columns_measurement
 from eds_scikit.biology.utils.process_measurement import filter_measurement_valid, filter_measurement_by_date, tag_measurement_anomaly, normalize_unit
-from eds_scikit.biology.utils.process_measurement import ConceptsSet
+from eds_scikit.biology.utils.process_concepts import ConceptsSet
 from eds_scikit.biology.utils.prepare_relationship import prepare_biology_relationship_table
 from eds_scikit.io.settings import mapping
 from eds_scikit.utils.typing import Data, DataFrame
