+++ eds_scikit/biology/viz/__init__.py
@@ -1,3 +1,2 @@
-from .plot import plot_concepts_set
-from .aggregate import aggregate_concepts_set
 from .wrapper import plot_biology_summary
+from .stats_summary import measurement_values_summary
