+++ eds_scikit/datasets/elements.csv
@@ -1,7 +1,7 @@
 unit_a,unit_b,conversion,element
 mol,g,113.12,Créatinine
-g,mol,0.00884016973125884,Créatinine
+g,mol,0.0088401697312588,Créatinine
 mol,g,180.156,Glucose
-g,mol,0.005550744909966918,Glucose
+g,mol,0.0055507449099669,Glucose
 mol,g,64500.0,Hémoglobine
 g,mol,1.5503875968992248e-05,Hémoglobine
